Fixed bug causing duplicates in output alignments
The issue was in filter_by_marker_nosplit.sh. The output *.filteredList needs to
have no duplicate entires, but sort -u or sort | uniq were too resource heavy. A
simple awk solution was used since the input to the awk command (line 130) was
assumed to be sorted. It turned out to not be sorted, so this was introducing
occasional duplicates. The solution was a simple script (src/lenFiltUniq.py)
which keeps track of the read ids already seen using a set. This has been tested
on HG002 v2.7 chrX and chrY primary alignments. Output is the same as before with
the following two differences: (1) the duplicates are gone and (2) the
*.filteredList files are in a different order (which is fine because they ended
up not being sorted before anyway). The alignments ($target.markersandlength.bam
and $target.markersandlength.paf) are still in the same order (sans the
duplicates, of course). Clearly, src/SubFile.java can use the filtered list
(*.filteredList) contents in any order while retaining the order in the sam file
it is subsetting.